[CAS] Update CAS implementation according to upstream version
Update downstream code to match upstreamed version.
